SourcePos: Remove from ispc.h
