+++ CMakeLists.txt
@@ -285,6 +285,8 @@ target_sources(${PROJECT_NAME}
         "src/expr.h"
         "src/func.cpp"
         "src/func.h"
+        "src/globals.h"
+        "src/globals.cpp"
         "src/ispc.cpp"
         "src/ispc.h"
         "src/ispc_version.h"
@@ -295,6 +297,8 @@ target_sources(${PROJECT_NAME}
         "src/module.h"
         "src/opt.cpp"
         "src/opt.h"
+        "src/source_pos.h"
+        "src/source_pos.cpp"
         "src/stmt.cpp"
         "src/stmt.h"
         "src/sym.cpp"
