+++ src/bitcode_lib.h
@@ -67,8 +67,8 @@ class BitcodeLib {
 
     BitcodeLibType getType() const;
     const unsigned char *getLib() const;
-    const size_t getSize() const;
-    const TargetOS getOS() const;
-    const Arch getArch() const;
-    const ISPCTarget getISPCTarget() const;
+    size_t getSize() const;
+    TargetOS getOS() const;
+    Arch getArch() const;
+    ISPCTarget getISPCTarget() const;
 };
