+++ src/globals.h
@@ -1,0 +1,13 @@
+#pragma once
+
+class Module;
+
+struct Globals;
+
+namespace ispc {
+
+Module *getModule();
+
+Globals *getGlobals();
+
+} // namespace ispc
