+++ src/ispc.h
@@ -63,6 +63,14 @@
 #include <string>
 #include <vector>
 
+namespace ispc {
+
+struct SourcePos;
+
+} // namespace ispc
+
+using ispc::SourcePos;
+
 /** @def ISPC_MAX_NVEC maximum vector size of any of the compliation
     targets.
  */
@@ -113,38 +121,6 @@ class Type;
 struct VariableDeclaration;
 
 enum StorageClass { SC_NONE, SC_EXTERN, SC_STATIC, SC_TYPEDEF, SC_EXTERN_C };
-
-/** @brief Representation of a range of positions in a source file.
-
-    This class represents a range of characters in a source file
-    (e.g. those that span a token's definition), from starting line and
-    column to ending line and column.  (These values are tracked by the
-    lexing code).  Both lines and columns are counted starting from one.
- */
-struct SourcePos {
-    SourcePos(const char *n = NULL, int fl = 0, int fc = 0, int ll = 0, int lc = 0);
-
-    const char *name;
-    int first_line;
-    int first_column;
-    int last_line;
-    int last_column;
-
-    /** Prints the filename and line/column range to standard output. */
-    void Print() const;
-
-    /** Returns a LLVM DIFile object that represents the SourcePos's file */
-    llvm::DIFile *GetDIFile() const;
-
-    /** Returns a LLVM DINamespace object that represents 'ispc' namespace. */
-    llvm::DINamespace *GetDINamespace() const;
-
-    bool operator==(const SourcePos &p2) const;
-};
-
-/** Returns a SourcePos that encompasses the extent of both of the given
-    extents. */
-SourcePos Union(const SourcePos &p1, const SourcePos &p2);
 
 /** @brief Structure that defines a compilation target
