+++ src/target_enums.cpp
@@ -39,6 +39,7 @@
 
 #include "ispc.h"
 #include "util.h"
+#include "source_pos.h"
 
 #include <cstring>
